fix: prevent reconfiguring the same device
This change prevents reconfiguration by breaking out of the loop.

BREAKING CHANGE: Due to having only a single device connected is the
more likely scenario, the prevention is fairly simple in the sense that
it only checks if any changes have been made other than checking changes
per device, therefore stopping after configuring the first device it
finds.

To use it with more devices, simply follow the directive in L#72 of
WheelSupports.cpp which is commenting out the break statement.